% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actionLink_oc.R
\name{actionLink_oc}
\alias{actionLink_oc}
\title{Create a shiny actionLink that triggers an off canvas}
\usage{
actionLink_oc(id_oc, ...)
}
\arguments{
\item{id_oc}{id of the off canvas to trigger}

\item{...}{Same parameters as shiny::actionButton(...)}
}
\value{
A shiny actionLink that triggers a particular off canvas
}
\description{
Create a shiny actionLink that triggers an off canvas
}
\examples{
actionLink_oc(
inputId = "ok",
label = "Trigger",
id_oc = "id_oc",
class="text-danger"
)
}
