% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_oc.R
\name{add_oc}
\alias{add_oc}
\title{Create a BS5 off canvas menu}
\usage{
add_oc(
  id = NULL,
  body = NULL,
  position = c("top", "bottom", "start", "end"),
  scroll = FALSE,
  backdrop = TRUE,
  header = TRUE,
  title = NULL,
  close_btn = TRUE,
  class_btn = NULL,
  class_oc = NULL,
  class_header = NULL,
  class_body = NULL
)
}
\arguments{
\item{id}{id of the off canvas menu}

\item{body}{HTML body of the off canvas menu}

\item{position}{position of the off-canvas (top, bottom, start or end)}

\item{scroll}{allow body scrolling (TRUE or FALSE)}

\item{backdrop}{allow body backdrop (TRUE or FALSE)}

\item{header}{Has the off canvas a header ? (TRUE or FALSE)}

\item{title}{HTML title of the off canvas menu}

\item{close_btn}{close button in the header ? (TRUE or FALSE)}

\item{class_btn}{Additional HTML class for the close button}

\item{class_oc}{Additional HTML class for the off canvas}

\item{class_header}{Additional HTML class for the off canvas title}

\item{class_body}{Additional HTML class for the off canvas body}
}
\value{
a BS5 off canvas menu
}
\description{
Create a BS5 off canvas menu
}
\examples{
add_oc(
id = "id_oc",
header = TRUE,
title = "Title !",
body ="Body !",
position = "end",
class_header = "bg-primary",
class_body = "bg-light",
class_btn = "btn-success",
class_oc ="border border-primary",
scroll = TRUE,
backdrop = TRUE,
close_btn = TRUE
)
}
