% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\name{add_rownames}
\alias{add_rownames}
\title{Add a new column with row names}
\usage{
add_rownames(.self, var = "rowname")
}
\arguments{
\item{.self}{Data frame}

\item{var}{Optional name of column}
}
\value{
Data frame
}
\description{
Add a new column with row names
}
\seealso{
Other column manipulations: \code{\link{define}},
  \code{\link{define_}}; \code{\link{rename}},
  \code{\link{rename_}}; \code{\link{select}},
  \code{\link{select_}}; \code{\link{undefine}},
  \code{\link{undefine_}}, \code{\link{unselect}},
  \code{\link{unselect_}}
}

