% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{between}
\alias{between}
\title{Tests whether elements of a vector lie between two values (inclusively)}
\usage{
between(x, left, right)
}
\arguments{
\item{x}{Values to test}

\item{left}{Left value}

\item{right}{Right value}
}
\value{
x >= left & x <= right
}
\description{
Tests whether elements of a vector lie between two values (inclusively)
}
\examples{
between (5, 1, 10)
between (11, 1, 10)
between (-5, -10, 0)
}
\seealso{
Other utility functions: \code{\link{cumall}},
  \code{\link{cumany}}, \code{\link{cummean}},
  \code{\link{distribute}}, \code{\link{first}},
  \code{\link{lag}}, \code{\link{last}},
  \code{\link{lead}}, \code{\link{n_distinct}},
  \code{\link{n_groups}}, \code{\link{nth}},
  \code{\link{n}}
}

