% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{bind_variables}
\alias{bind_variables}
\title{Bind parallel data frame variables to an environment}
\usage{
bind_variables(dat, envir)
}
\arguments{
\item{dat}{Data frame}

\item{envir}{Environment}
}
\description{
Bind parallel data frame variables to an environment
}

