% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button_app.R
\name{button_app}
\alias{button_app}
\title{Creation of a bootstrap 5 button \strong{work in progress}}
\usage{
button_app(
  id,
  color = "primary",
  href = NULL,
  outline = FALSE,
  shadow = TRUE,
  oc = FALSE,
  id_oc = NULL,
  text = NULL,
  icon = NULL
)
}
\arguments{
\item{id}{button id}

\item{color}{button bootstrap color (primary, secondary, info, success, warning, danger, dark, light)}

\item{href}{link to an external website (default = NULL)}

\item{outline}{outline button ? (TRUE/FALSE)}

\item{shadow}{shadow around button ? (TRUE/FALSE)}

\item{oc}{Trigger off canvas ? (TRUE/FALSE)}

\item{id_oc}{if oc is TRUE id of the oc to trigger}

\item{text}{Text label (default = NULL)}

\item{icon}{Icon label (default = NULL)}
}
\value{
a bootstrap 5 button
}
\description{
Creation of a bootstrap 5 button \strong{work in progress}
}
