% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{clsna}
\alias{.clsna}
\alias{clsna}
\title{Returns NA as a cluster object}
\usage{
.clsna()
}
\value{
NA with class set to SOCKcluster
}
\description{
This is a horrible kludge needed so copy() doesn't complain about
assigning NULL to a field when there's no cluster, i.e. for the slave nodes
}
\keyword{internal}

