% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_multiplication.R
\name{coord_line}
\alias{coord_line}
\title{calculate start and end coordinates of segments}
\usage{
coord_line(nb_vertice, modulo, table)
}
\arguments{
\item{nb_vertice}{number of polygon segments}

\item{modulo}{number of points}

\item{table}{desired multiplication table}
}
\value{
return coordinates of segments
}
\description{
calculate start and end coordinates of segments
}
\examples{
library(dplyr)
segment <- coord_line(nb_vertice = 5, modulo = 100,table = 2)
}
