% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_multiplication.R
\name{coord_point_pg}
\alias{coord_point_pg}
\title{calculate coordinates of points regularly distributed on polygon segments}
\usage{
coord_point_pg(nb_vertice, modulo)
}
\arguments{
\item{nb_vertice}{number of polygon segments}

\item{modulo}{number of points}
}
\value{
return coordinates of points
}
\description{
calculate coordinates of points regularly distributed on polygon segments
}
\examples{
library(dplyr)
poly <- coord_pg(3)
point <- coord_point_pg(nb_vertice=3,modulo=20)
}
