% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{cumall}
\alias{cumall}
\title{Cumulative all}
\usage{
cumall(x)
}
\arguments{
\item{x}{Values}
}
\value{
Cumulative all of x
}
\description{
Cumulative all
}
\seealso{
Other utility functions: \code{\link{between}};
  \code{\link{cumany}}; \code{\link{cummean}};
  \code{\link{distribute}}; \code{\link{first}};
  \code{\link{lag}}; \code{\link{last}};
  \code{\link{lead}}; \code{\link{n_distinct}};
  \code{\link{n_groups}}; \code{\link{nth}};
  \code{\link{n}}
}

