% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{cumany}
\alias{cumany}
\title{Cumulative any}
\usage{
cumany(x)
}
\arguments{
\item{x}{Values}
}
\value{
Cumulative any of x
}
\description{
Cumulative any
}
\seealso{
Other utility functions: \code{\link{between}};
  \code{\link{cumall}}; \code{\link{cummean}};
  \code{\link{distribute}}; \code{\link{first}};
  \code{\link{lag}}; \code{\link{last}};
  \code{\link{lead}}; \code{\link{n_distinct}};
  \code{\link{n_groups}}; \code{\link{nth}};
  \code{\link{n}}
}

