% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{cummean}
\alias{cummean}
\title{Cumulative mean}
\usage{
cummean(x)
}
\arguments{
\item{x}{Values to obtain cumulative mean of}
}
\value{
Cumulative mean of x
}
\description{
Cumulative mean
}
\seealso{
Other utility functions: \code{\link{between}};
  \code{\link{cumall}}; \code{\link{cumany}};
  \code{\link{distribute}}; \code{\link{first}};
  \code{\link{lag}}; \code{\link{last}};
  \code{\link{lead}}; \code{\link{n_distinct}};
  \code{\link{n_groups}}; \code{\link{nth}};
  \code{\link{n}}
}

