% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{define}
\alias{define}
\alias{define_}
\title{Define new variables}
\usage{
define(.self, ...)

define_(.self, ..., .dots)
}
\arguments{
\item{.self}{Data frame}

\item{...}{Names of new variables}

\item{.dots}{Workaround for non-standard evaluation}
}
\description{
Define new variables
}
\section{Functions}{
\itemize{
\item \code{define_}: 
}}

