% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{desc}
\alias{desc}
\title{Arrange specified column in descending order}
\usage{
desc(var)
}
\arguments{
\item{var}{Variable to arrange in descending order}
}
\description{
Arrange specified column in descending order
}
\examples{
\donttest{
dat <- Multiplyr (x=1:100, cl=2)
dat \%>\% arrange(desc(x))
dat \%>\% shutdown()
}
}
\seealso{
\code{\link{arrange}}
}

