% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{distinct}
\alias{distinct}
\alias{distinct_}
\title{Select unique rows or unique combinations of variables}
\usage{
distinct(.self, ..., auto_compact = NULL)

distinct_(.self, ..., .dots, auto_compact = NULL)
}
\arguments{
\item{.self}{Data frame}

\item{...}{Additional parameters}

\item{auto_compact}{Compact data after operation}

\item{.dots}{Workaround for non-standard evaluation}
}
\description{
Select unique rows or unique combinations of variables
}
\section{Functions}{
\itemize{
\item \code{distinct_}: 
}}

