% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{distribute}
\alias{distribute}
\title{Calculations for how to distribute x items over N nodes}
\usage{
distribute(x, N)
}
\arguments{
\item{x}{Number of items or a vector of group sizes}

\item{N}{Number of nodes}
}
\description{
Calculations for how to distribute x items over N nodes
}

