% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{.dots2names}
\alias{.dots2names}
\title{Extract names from a lazy_dots object (internal)}
\usage{
.dots2names(dots)
}
\arguments{
\item{dots}{Lazy dots object}
}
\value{
Vector of names
}
\description{
This will take a lazy_dots object and will extract all the names from it
}
\examples{
f <- function (...) { lazyeval::lazy_dots (...) }
dots <- f(x, y=z)
.dots2names (dots)
}
\keyword{internal}

