% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{dotscapture}
\alias{dotscapture}
\title{Capture ... for later evaluation}
\usage{
dotscapture(...)
}
\arguments{
\item{...}{Dots from function}
}
\value{
Dots object for use with \code{\link{dotseval}}
}
\description{
Capture ... for later evaluation
}
\keyword{internal}

