% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{dotscombine}
\alias{dotscombine}
\title{Combine explicit and implicit dots}
\usage{
dotscombine(dots, ...)
}
\value{
Dots object for use with \code{\link{dotseval}}
}
\description{
Combine explicit and implicit dots
}
\keyword{internal}

