% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{dotseval}
\alias{dotseval}
\title{Evaluate previously captured dots}
\usage{
dotseval(dots, env)
}
\arguments{
\item{dots}{Captured ... from \code{\link{dotscapture}}}

\item{env}{Environment to evaluate in}
}
\value{
Results of evaluating expression
}
\description{
Evaluate previously captured dots
}
\keyword{internal}

