% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{dotsname}
\alias{dotsname}
\title{Ensure captured dots are all named}
\usage{
dotsname(dots)
}
\arguments{
\item{dots}{Captured dots}
}
\value{
Captured dots, all named
}
\description{
Ensure captured dots are all named
}
\keyword{internal}

