% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{dotsname1}
\alias{dotsname1}
\title{Name an expression (called by dotsname)}
\usage{
dotsname1(expr)
}
\arguments{
\item{expr}{Expression}
}
\value{
Name of expression
}
\description{
Name an expression (called by dotsname)
}
\keyword{internal}

