% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwld_button_app.R
\name{dwld_button_app}
\alias{dwld_button_app}
\title{wrapper of downloadButton to suppress btn-default}
\usage{
dwld_button_app(...)
}
\arguments{
\item{...}{parameters of downloadButton}
}
\value{
a download button without btn-default
}
\description{
wrapper of downloadButton to suppress btn-default
}
