% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{fast_filter}
\alias{fast_filter}
\alias{fast_filter_}
\title{Faster form of usual filter function, but has restrictions}
\usage{
fast_filter(.data, ...)

fast_filter_(.data, ..., .dots)
}
\arguments{
\item{.data}{Data frame}

\item{...}{Filtering expressions}

\item{.dots}{Workaround for non-standard evaluation}
}
\description{
Faster form of usual filter function, but has restrictions
}
\section{Functions}{
\itemize{
\item \code{fast_filter_}: 
}}

