% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{filter}
\alias{filter}
\alias{filter_}
\title{Filter data}
\usage{
filter(.self, ...)

filter_(.self, ..., .dots, auto_compact = NULL)
}
\arguments{
\item{...}{Additional parameters}

\item{.dots}{Workaround for non-standard evaluation}

\item{.data}{Data frame}
}
\description{
Filter data
}
\section{Functions}{
\itemize{
\item \code{filter_}: 
}}

