% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{first}
\alias{first}
\title{Returns first value in vector}
\usage{
first(x, default = NA)
}
\arguments{
\item{x}{Vector}

\item{default}{Value to pad missing data with}
}
\value{
First value in x
}
\description{
Returns first value in vector
}
\examples{
first (1:10)
first (numeric(0), default=123)
}
\seealso{
Other utility functions: \code{\link{between}},
  \code{\link{cumall}}, \code{\link{cumany}},
  \code{\link{cummean}}, \code{\link{distribute}},
  \code{\link{lag}}, \code{\link{last}},
  \code{\link{lead}}, \code{\link{n_distinct}},
  \code{\link{n_groups}}, \code{\link{nth}},
  \code{\link{n}}
}

