% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_multiplication.R
\name{graph_line}
\alias{graph_line}
\title{Graphic representation of a particular multiplication table}
\usage{
graph_line(
  nb_vertice,
  modulo,
  table,
  zoom = 1,
  bgcolor = "transparent",
  curvature,
  angle,
  alpha,
  colour
)
}
\arguments{
\item{nb_vertice}{number of polygon segments}

\item{modulo}{number of points}

\item{table}{desired multiplication table}

\item{zoom}{level of zoom}

\item{bgcolor}{Background color}

\item{curvature}{Level of segment curvature}

\item{angle}{Angle of segment curvature}

\item{alpha}{Alpha level of segment color}

\item{colour}{Segment color}
}
\value{
a graphic of the multiplication table
}
\description{
Graphic representation of a particular multiplication table
}
\examples{
library(dplyr)
library(ggplot2)
graph_line(nb_vertice = 5,
modulo = 20000,
table = 3,
alpha = 0.05,
curvature = 0,
angle = 0,
colour = "#2c3e50")
}
