% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{group_restrict}
\alias{group_restrict}
\title{Return a parallel data frame mapped to a particular group}
\usage{
group_restrict(dat, group)
}
\arguments{
\item{dat}{Data frame}

\item{group}{Group ID}
}
\description{
Return a parallel data frame mapped to a particular group
}

