% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\name{group_sizes}
\alias{group_sizes}
\title{Return size of groups}
\usage{
group_sizes(.self)
}
\arguments{
\item{.self}{Data frame}
}
\value{
Group sizes
}
\description{
This function is used to find the size of groups in a Multiplyr data frame
}
\examples{
\donttest{
dat <- Multiplyr (x=1:100, G=rep(c("A", "B", "C", "D"), length.out=100))
dat \%>\% group_by (G)
group_sizes (dat)
dat \%>\% shutdown()
}
}

