% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{internal}
\alias{.dots2names}
\alias{.filter_range}
\alias{.p}
\alias{internal}
\alias{no.strings.attached}
\title{Internal functions}
\usage{
.p(...)

.dots2names(dots)

no.strings.attached(x)

.filter_range(x, filtercol, start, end)
}
\description{
Internal functions
}
\keyword{internal}

