% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{last}
\alias{last}
\title{Returns last value in vector}
\usage{
last(x, default = NA)
}
\arguments{
\item{x}{Vector}

\item{default}{Value to pad missing data with}
}
\value{
Last value in x
}
\description{
Returns last value in vector
}
\examples{
last (1:10)
}
\seealso{
Other utility functions: \code{\link{between}},
  \code{\link{cumall}}, \code{\link{cumany}},
  \code{\link{cummean}}, \code{\link{distribute}},
  \code{\link{first}}, \code{\link{lag}},
  \code{\link{lead}}, \code{\link{n_distinct}},
  \code{\link{n_groups}}, \code{\link{nth}},
  \code{\link{n}}
}

