% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Multiplyr-class}
\alias{Multiplyr}
\alias{Multiplyr-class}
\title{Create new parallel data frame}
\arguments{
\item{cl}{Cluster object, number of nodes or NULL (default)}

\item{alloc}{Allocate additional space}
}
\description{
Create new parallel data frame
}

