% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{mutate}
\alias{mutate}
\alias{mutate_}
\title{Change values of existing variables (and create new ones)}
\usage{
mutate(.self, ...)

mutate_(.self, ..., .dots)
}
\arguments{
\item{.self}{Data frame}

\item{...}{Additional parameters}

\item{.dots}{Workaround for non-standard evaluation}
}
\description{
Change values of existing variables (and create new ones)
}
\section{Functions}{
\itemize{
\item \code{mutate_}: 
}}

