% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{n}
\alias{n}
\title{Number of items in current group}
\usage{
n()
}
\value{
Number of items in current group (or node if ungrouped)
}
\description{
Number of items in current group
}
\seealso{
Other utility functions: \code{\link{between}};
  \code{\link{cumall}}; \code{\link{cumany}};
  \code{\link{cummean}}; \code{\link{distribute}};
  \code{\link{first}}; \code{\link{lag}};
  \code{\link{last}}; \code{\link{lead}};
  \code{\link{n_distinct}}; \code{\link{n_groups}};
  \code{\link{nth}}
}

