% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{n_distinct}
\alias{n_distinct}
\title{Return the number of unique values}
\usage{
n_distinct(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Number of unique values of x
}
\description{
Return the number of unique values
}
\examples{
n_distinct (c(1, 2, 3, 3, 3, 4, 5))
}
\seealso{
Other utility functions: \code{\link{between}},
  \code{\link{cumall}}, \code{\link{cumany}},
  \code{\link{cummean}}, \code{\link{distribute}},
  \code{\link{first}}, \code{\link{lag}},
  \code{\link{last}}, \code{\link{lead}},
  \code{\link{n_groups}}, \code{\link{nth}},
  \code{\link{n}}
}

