% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\name{n_groups}
\alias{n_groups}
\title{Return number of groups}
\usage{
n_groups(.self)
}
\arguments{
\item{.self}{Data frame}
}
\value{
Number of groups in data frame
}
\description{
Return number of groups
}
\seealso{
Other utility functions: \code{\link{between}};
  \code{\link{cumall}}; \code{\link{cumany}};
  \code{\link{cummean}}; \code{\link{distribute}};
  \code{\link{first}}; \code{\link{lag}};
  \code{\link{last}}; \code{\link{lead}};
  \code{\link{n_distinct}}; \code{\link{nth}};
  \code{\link{n}}
}

