% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{nonunique}
\alias{nonunique}
\title{Returns values of x that are non-unique}
\usage{
nonunique(x)
}
\description{
Returns values of x that are non-unique
}
\examples{
\donttest{
nonunique (c(1, 1, 1, 2, 3))
nonunique (c(1, 2, 3, 3, 4, 4))
}
}
\keyword{internal}

