% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{nth}
\alias{nth}
\title{Return the nth item from a vector}
\usage{
nth(x, n, default = NA)
}
\arguments{
\item{x}{Vector}

\item{n}{The n in nth}

\item{default}{Value to pad missing data with}
}
\value{
nth item from vector
}
\description{
Return the nth item from a vector
}
\examples{
nth (1:10, 5)
nth (1:10, 11, default=123)
}
\seealso{
Other utility functions: \code{\link{between}},
  \code{\link{cumall}}, \code{\link{cumany}},
  \code{\link{cummean}}, \code{\link{distribute}},
  \code{\link{first}}, \code{\link{lag}},
  \code{\link{last}}, \code{\link{lead}},
  \code{\link{n_distinct}}, \code{\link{n_groups}},
  \code{\link{n}}
}

