% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{.p}
\alias{.p}
\title{Concatenate (internal)}
\usage{
.p(...)
}
\arguments{
\item{...}{Strings to be concatenated}
}
\value{
Concetenated string
}
\description{
Shorthand for concatenating strings together
}
\examples{
cat (.p("hello ", "world!"))
}
\keyword{internal}

