% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\name{partition}
\alias{partition}
\title{Partition data evenly amongst cluster nodes}
\usage{
partition(.data, max.row = NULL)
}
\arguments{
\item{.data}{Data frame}

\item{max.row}{Partition only these number of rows. NULL (default) or 0
will partition the entire data set}
}
\description{
Partition data evenly amongst cluster nodes
}

