% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\name{partition_even}
\alias{partition_even}
\title{Partition data evenly amongst cluster nodes}
\usage{
partition_even(.self)
}
\arguments{
\item{.self}{Data frame}
}
\value{
Data frame
}
\description{
This function results in data being repartitioned evenly across cluster nodes,
ignoring any grouping variables.
}
\seealso{
Other cluster functions: \code{\link{partition_group}},
  \code{\link{partition_group_}}; \code{\link{shutdown}}
}

