% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{reduce}
\alias{reduce}
\alias{reduce_}
\title{Summarise data (with local reduction)}
\usage{
reduce(.self, ...)

reduce_(.self, ..., .dots, auto_compact = NULL)
}
\description{
\code{summarise} is used to summarise data on each node: \code{reduce} is then used to ensure that there's one overall summary
}
\details{
When data have not been grouped, calling \code{summary(...)} will result
in each node summarising the data it has available. This means that if
there are 3 nodes in the cluster, then there will be 3 summary values.
\code{reduce} is used to bring all those together to a single value.
}
\examples{
\donttest{
dat <- Multiplyr (x = 1:100)
dat \%>\% summarise (N = length(x))

dat <- Multiplyr (x = 1:100)
dat \%>\% summarise (N = length(x)) \%>\% reduce(N = sum(N))
}
}

