% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\name{regroup}
\alias{groupwise}
\alias{regroup}
\title{Return to grouped data}
\usage{
regroup(.data)

groupwise(.data)
}
\arguments{
\item{.data}{Data frame}
}
\description{
Return to grouped data
}
\section{Functions}{
\itemize{
\item \code{groupwise}: 
}}

