% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{rename}
\alias{rename}
\alias{rename_}
\title{Rename variables}
\usage{
rename(.self, ...)

rename_(.self, ..., .dots)
}
\arguments{
\item{.self}{Data frame}

\item{...}{Additional parameters}

\item{.dots}{Workaround for non-standard evaluation}
}
\description{
Rename variables
}
\section{Functions}{
\itemize{
\item \code{rename_}: 
}}

