% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app_multiplyr.R
\name{run_app_multiplyr}
\alias{run_app_multiplyr}
\title{Run the shiny multiplyr app}
\usage{
run_app_multiplyr()
}
\value{
Launching of the shiny multiplyr app
}
\description{
Run the shiny multiplyr app
}
