% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{select}
\alias{select}
\alias{select_}
\title{Retain only specified variables}
\usage{
select(.data, ...)

select_(.data, ..., .dots)
}
\arguments{
\item{.data}{Data frame}

\item{...}{Additional parameters}

\item{.dots}{Workaround for non-standard evaluation}
}
\description{
Retain only specified variables
}
\section{Functions}{
\itemize{
\item \code{select_}: 
}}

