% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\name{shutdown}
\alias{shutdown}
\title{Shutdown running cluster}
\usage{
shutdown(.self)
}
\arguments{
\item{.self}{Data frame}
}
\description{
Theoretically a Multiplyr data frame will have its cluster implicitly
shutdown when R's garbage collection kicks in. This function exists to
execute it explicitly. This does not affect any of the data.
}
\seealso{
Other cluster functions: \code{\link{partition_even}};
  \code{\link{partition_group}},
  \code{\link{partition_group_}}
}

