% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\name{slice}
\alias{slice}
\title{Select rows by position}
\usage{
slice(.self, rows = NULL, start = NULL, end = NULL, each = FALSE,
  auto_compact = NULL)
}
\arguments{
\item{.self}{Data frame}

\item{rows}{Rows to select}

\item{start}{Start of range of rows}

\item{end}{End of range of rows}

\item{each}{Apply slice to each cluster/group}

\item{auto_compact}{Compact data}
}
\description{
Select rows by position
}

