% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sm_desc_comp}
\alias{sm_desc_comp}
\title{Returns big.matrix descriptor offset by 1 (for row by row comparisons)}
\usage{
sm_desc_comp(.self, start)
}
\arguments{
\item{.self}{Data frame}

\item{start}{1 or 2}
}
\value{
big.matrix.descriptor subset to 1:last-1 or 2:last
}
\description{
Returns big.matrix descriptor offset by 1 (for row by row comparisons)
}
\keyword{internal}

