% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sm_desc_group}
\alias{sm_desc_group}
\title{Returns a big.matrix descriptor for a particular group ID}
\usage{
sm_desc_group(.self, group_id)
}
\arguments{
\item{.self}{Data frame}

\item{group_id}{Group ID (from groupcol)}
}
\value{
big.matrix.descriptor that would restrict to this group
}
\description{
Returns a big.matrix descriptor for a particular group ID
}
\keyword{internal}

