% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sm_desc_subset}
\alias{sm_desc_subset}
\title{Returns big.matrix descriptor limited to particular start/end row}
\usage{
sm_desc_subset(.self, first, last)
}
\arguments{
\item{.self}{Data frame}

\item{first}{First row}

\item{last}{Last row}
}
\value{
big.matrix.descriptor that would restrict to this group
}
\description{
Returns big.matrix descriptor limited to particular start/end row
}
\keyword{internal}

