% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sm_desc_update}
\alias{sm_desc_update}
\title{Update description of a big.matrix after a row subset (internal)}
\usage{
sm_desc_update(desc, first, last)
}
\arguments{
\item{desc}{Existing big.matrix.descriptor}

\item{first}{First relative row of that matrix}

\item{last}{Last relative row of that matrix}
}
\value{
New descriptor
}
\description{
Generating a new big.matrix.descriptor or doing sub.big.matrix on something
that's not a descriptor is slow. This method exists to effectively create
the descriptor that describe(new.sub.big.matrix) would do, but in a fraction
of the time.
}
\keyword{internal}

