% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{reduce}
\alias{reduce}
\alias{reduce_}
\alias{summarise}
\alias{summarise_}
\title{Summarise data locally with reduction}
\usage{
reduce(.self, ...)

summarise(.self, ..., auto_compact = NULL)

reduce_(.self, ..., .dots, auto_compact = NULL)

summarise_(.self, ..., .dots, auto_compact = NULL)
}
\arguments{
\item{.self}{Data frame}

\item{...}{Additional parameters}

\item{auto_compact}{Compact data after operation}

\item{.dots}{Workaround for non-standard evaluation}
}
\description{
Summarise data locally with reduction

Summarise data
}
\section{Functions}{
\itemize{
\item \code{reduce}: 

\item \code{reduce_}: 

\item \code{summarise_}: 
}}

