% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bs_color.R
\name{test_bs_color}
\alias{test_bs_color}
\title{Test the bootstrap validity of a color}
\usage{
test_bs_color(color, transparent = FALSE, default = FALSE, bw = TRUE)
}
\arguments{
\item{color}{color to test}

\item{transparent}{Is transparent color authorized ?}

\item{default}{Is default color authorized ?}

\item{bw}{Is black ans white authorized ?}
}
\value{
A error message if the color is invalid
}
\description{
Test the bootstrap validity of a color
}
\examples{
test_bs_color(
color = "default" 
)
test_bs_color(
color = "default",
default = TRUE
)
}
