% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{test_transition}
\alias{test_transition}
\title{Test for grouping transition (internal)}
\usage{
test_transition(.self, cols, rows)
}
\arguments{
\item{.self}{Parallel data frame}

\item{cols}{Columns to test}

\item{rows}{Rows to test}
}
\value{
Logical vector of which rows have a transition (first element is always TRUE)
}
\description{
This algorithm tests specific rows for a transition. Each cluster node will
test its own subset for grouping transitions, but the transition between the
last row of one cluster and the first of the next needs to be tested on the
master node.
}
\keyword{internal}

