% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{undefine}
\alias{undefine}
\alias{undefine_}
\alias{unselect}
\alias{unselect_}
\title{Delete variables}
\usage{
undefine(.data, ...)

unselect(.data, ...)

undefine_(.data, ..., .dots)

unselect_(.data, ..., .dots)
}
\arguments{
\item{.data}{Data frame}

\item{...}{Additional parameters}

\item{.dots}{Workaround for non-standard evaluation}
}
\description{
Delete variables
}
\section{Functions}{
\itemize{
\item \code{unselect}: 

\item \code{undefine_}: 

\item \code{unselect_}: 
}}

