% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\name{ungroup}
\alias{rowwise}
\alias{ungroup}
\title{Return data to non-grouped}
\usage{
ungroup(.data)

rowwise(.data)
}
\arguments{
\item{.data}{Data frame}

\item{...}{Additional parameters}

\item{.dots}{Workaround for non-standard evaluation}
}
\description{
Return data to non-grouped
}
\section{Functions}{
\itemize{
\item \code{rowwise}: 
}}

