% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{warn.suboptimal}
\alias{warn.suboptimal}
\title{Display a warning if a condition is sub-optimal}
\usage{
warn.suboptimal(expr, warn)
}
\arguments{
\item{expr}{Expression}

\item{warn}{Warning text}
}
\description{
Display a warning if a condition is sub-optimal
}

