% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\name{within_group}
\alias{within_group}
\title{Execute code within a group}
\usage{
within_group(.data, .expr)
}
\arguments{
\item{.data}{Data frame}

\item{.expr}{Code to execute}
}
\description{
Execute code within a group
}

