% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\name{within_node}
\alias{within_node}
\title{Execute code within a node}
\usage{
within_node(.data, .expr)
}
\arguments{
\item{.data}{Data frame}

\item{.expr}{Code to execute}
}
\description{
Execute code within a node
}

