% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{arrange}
\alias{arrange}
\alias{arrange_}
\title{Sort data frame}
\usage{
arrange(.self, ...)

arrange_(.self, ..., .dots)
}
\arguments{
\item{...}{Variables to sort by}

\item{.dots}{Workaround for non-standard evaluation}

\item{.data}{Data frame}
}
\description{
Sort data frame
}
\section{Functions}{
\itemize{
\item \code{arrange_}: 
}}

