% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{define}
\alias{define}
\alias{define_}
\title{Define new variables}
\usage{
define(.self, ...)

define_(.self, ..., .dots)
}
\arguments{
\item{.self}{Data frame}

\item{...}{Names of new variables}

\item{.dots}{Workaround for non-standard evaluation}
}
\description{
This function is used to create new columns in a data frame. Specifying
just the name will create a new numeric column. If specified in the form
of \code{var=template}, then a new column named \code{var} will be
created using \code{template} as a template. This is useful for creating
new factors.
}
\section{Functions}{
\itemize{
\item \code{define_}: 
}}

