% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{define}
\alias{define}
\alias{define_}
\title{Define new columns}
\usage{
define(.self, ...)

define_(.self, ..., .dots)
}
\arguments{
\item{.self}{Data frame}

\item{...}{Additional parameters}

\item{.dots}{Workaround for non-standard evaluation}
}
\value{
Data frame
}
\description{
Takes a list of column names and creates them. Optionally uses a template to
copy across factor/character mappings.
}
\details{
This function is used to create new columns in a data frame. Specifying
just the name will create a new numeric column. If specified in the form
of \code{var=template}, then a new column named \code{var} will be
created using \code{template} as a template. This is useful for creating
new factors.
}
\seealso{
Other column manipulations: \code{\link{add_rownames}},
  \code{\link{rename}}, \code{\link{select}},
  \code{\link{undefine}}
}

