% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{distinct}
\alias{distinct}
\alias{distinct_}
\title{Select unique rows or unique combinations of variables}
\usage{
distinct(.self, ...)

distinct_(.self, ..., .dots)
}
\arguments{
\item{...}{Additional parameters}

\item{.dots}{Workaround for non-standard evaluation}

\item{.data}{Data frame}
}
\description{
Select unique rows or unique combinations of variables
}
\section{Functions}{
\itemize{
\item \code{distinct_}: 
}}

