% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{dotscapture}
\alias{dotscapture}
\title{Capture ... for later evaluation}
\usage{
dotscapture(...)
}
\arguments{
\item{...}{Dots from function}
}
\value{
Dots object for use with \code{\link{dotseval}}
}
\description{
Capture ... for later evaluation
}
\examples{
\donttest{
f <- function (...) { dotscapture(...) }
x <- 123
dots <- f (x)
dotseval (dots, sys.frame())
}
}
\keyword{internal}

