% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{dotseval}
\alias{dotseval}
\title{Evaluate previously captured dots}
\usage{
dotseval(expr, envir)
}
\arguments{
\item{dots}{Captured ... from \code{\link{dotscapture}}}

\item{env}{Environment to evaluate in}
}
\value{
Results of evaluating expression
}
\description{
Evaluate previously captured dots
}
\examples{
\donttest{
f <- function (...) { dotscapture(...) }
x <- 123
dots <- f (x)
dotseval (dots, sys.frame())
}
}
\keyword{internal}

