% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastdf.R
\name{fastdf}
\alias{fastdf}
\title{Create new parallel data frame}
\usage{
fastdf(..., alloc = 1, cl = NULL)
}
\arguments{
\item{alloc}{Allocate additional space}

\item{.cl}{Cluster object, number of nodes or NULL (default)}
}
\description{
Create new parallel data frame
}

