% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{filter}
\alias{filter}
\alias{filter_}
\title{Filter data}
\usage{
filter(.data, ...)

filter_(.data, ..., .dots)
}
\arguments{
\item{.data}{Data frame}

\item{...}{Additional parameters}

\item{.dots}{Workaround for non-standard evaluation}
}
\description{
Filter data
}
\section{Functions}{
\itemize{
\item \code{filter_}: 
}}

