% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_multiplication.R
\name{graph_line}
\alias{graph_line}
\title{Graphic representation of a particular multiplication table}
\usage{
graph_line(
  nb_vertice,
  modulo,
  table,
  zoom = 1,
  bgcolor = "transparent",
  curvature,
  angle,
  alpha,
  colour,
  outer_glow = TRUE,
  outer_glow_color = "#18BC9C",
  outer_glow_sigma = 1,
  outer_glow_expand = 1,
  inner_glow = TRUE,
  inner_glow_color = "#18BC9C",
  inner_glow_sigma = 0.5,
  inner_glow_expand = 1
)
}
\arguments{
\item{nb_vertice}{number of polygon segments}

\item{modulo}{number of points}

\item{table}{desired multiplication table}

\item{zoom}{level of zoom}

\item{bgcolor}{Background color}

\item{curvature}{Level of segment curvature}

\item{angle}{Angle of segment curvature}

\item{alpha}{Alpha level of segment color}

\item{colour}{Segment color}

\item{outer_glow}{outer glow (TRUE/FALSE)}

\item{outer_glow_color}{outer glow color}

\item{outer_glow_sigma}{outer glow sigma}

\item{outer_glow_expand}{outer glow expand}

\item{inner_glow}{inner glow (TRUE/FALSE)}

\item{inner_glow_color}{inner glow color}

\item{inner_glow_sigma}{inner glow sigma}

\item{inner_glow_expand}{inner glow expand}
}
\value{
a graphic of the multiplication table
}
\description{
Graphic representation of a particular multiplication table
}
\examples{
library(dplyr)
library(ggplot2)
graph_line(nb_vertice = 5,
modulo = 20000,
table = 3,
alpha = 0.05,
curvature = 0,
angle = 0,
colour = "#2c3e50")
}
