% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{group_by}
\alias{group_by}
\alias{group_by_}
\title{Group data}
\usage{
group_by(.self, ...)

group_by_(.self, ..., .dots, .cols = NULL)
}
\arguments{
\item{.self}{Data frame}

\item{...}{Variables to sort by}

\item{.dots}{Workaround for non-standard evaluation}
}
\description{
Group data
}
\section{Functions}{
\itemize{
\item \code{group_by_}: 
}}

