% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{internal}
\alias{.dots2names}
\alias{.filter_range}
\alias{.p}
\alias{internal}
\alias{no.strings.attached}
\alias{sm_desc_update}
\alias{test_transition}
\title{Internal functions}
\usage{
.p(...)

.dots2names(dots)

no.strings.attached(x)

.filter_range(x, filtercol, start, end)

sm_desc_update(desc, first, last)

test_transition(.self, cols, rows)
}
\description{
Internal functions
}
\keyword{internal}

