% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{internal}
\alias{.dots2names}
\alias{.filter_rewrite}
\alias{.p}
\alias{.partition_all}
\alias{.sort.fastdf}
\alias{alloc_col}
\alias{ff_eq}
\alias{ff_expr}
\alias{ff_ge}
\alias{ff_gt}
\alias{ff_le}
\alias{ff_lt}
\alias{ff_mwhich}
\alias{ff_neq}
\alias{free_col}
\alias{internal}
\alias{no.strings.attached}
\alias{pad.cols}
\title{Internal functions}
\usage{
.p(...)

.dots2names(x, dots)

.sort.fastdf(x, decreasing = FALSE, dots = NULL, cols = NULL,
  with.group = FALSE)

.filter_rewrite(lazyobj)

ff_expr(var, expr)

ff_lt(cmp, expr)

ff_le(cmp, expr)

ff_gt(cmp, expr)

ff_ge(cmp, expr)

ff_neq(cmp, expr)

ff_eq(cmp, expr)

ff_mwhich(x, lazyobj)

alloc_col(x, name = ".tmp")

free_col(x, col)

.partition_all(.data, max.row = nrow(.data[[1]]))

no.strings.attached(x)

pad.cols(x, max.row = 10)
}
\description{
Internal functions
}
\keyword{internal}

