% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{lag}
\alias{lag}
\title{Offset x backwards by n}
\usage{
lag(x, n = 1, default = NA)
}
\arguments{
\item{x}{Vector}

\item{n}{Amount of offset by}

\item{default}{Value to pad missing data with}
}
\value{
Offset values
}
\description{
Offset x backwards by n
}
\examples{
lag (1:10)
lag (1:10, 2)
}
\seealso{
Other utility functions: \code{\link{between}},
  \code{\link{cumall}}, \code{\link{cumany}},
  \code{\link{cummean}}, \code{\link{distribute}},
  \code{\link{first}}, \code{\link{last}},
  \code{\link{lead}}, \code{\link{n_distinct}},
  \code{\link{n_groups}}, \code{\link{nth}},
  \code{\link{n}}
}

