% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class.R
\docType{methods}
\name{Multiplyr-methods}
\alias{Multiplyr-methods}
\alias{[,Multiplyr,ANY,ANY,ANY-method}
\alias{[<-,Multiplyr,ANY,ANY-method}
\alias{as.data.frame,Multiplyr-method}
\alias{dimnames,Multiplyr-method}
\alias{names,Multiplyr-method}
\alias{row.names,Multiplyr-method}
\title{Data access methods for Multiplyr}
\usage{
\S4method{[}{Multiplyr,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{Multiplyr,ANY,ANY}(x, i, j, ...) <- value

\S4method{as.data.frame}{Multiplyr}(x)

\S4method{dimnames}{Multiplyr}(x)

\S4method{names}{Multiplyr}(x)

\S4method{row.names}{Multiplyr}(x)
}
\arguments{
\item{x}{Data frame}

\item{i}{Rows}

\item{j}{Columns}

\item{...}{Additional parameters}

\item{drop}{Logical: whether to reduce a 1 column data frame result to a vector}

\item{value}{Value to set}
}
\description{
Data access methods for Multiplyr

Get data

Set data

Coerce Multiplyr object to be a standard data.frame

List containing row and column names

Column names

Row names
}

