% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{mutate}
\alias{mutate}
\alias{mutate_}
\title{Change values of existing variables (and create new ones)}
\usage{
mutate(.self, ...)

mutate_(.self, ..., .dots)
}
\arguments{
\item{.self}{Data frame}

\item{...}{Additional parameters}

\item{.dots}{Workaround for non-standard evaluation}
}
\value{
Data frame
}
\description{
This function is used to alter the data frame, without dropping any columns (unlike \code{transmute}, which drops any columns not explicitly specified)
}
\examples{
\donttest{
dat <- Multiplyr (x=1:100)
dat \%>\% mutate (x=x*2)
dat \%>\% shutdown()
}
}
\seealso{
Other data manipulations: \code{\link{nsa}},
  \code{\link{reduce}}, \code{\link{summarise}},
  \code{\link{transmute}}, \code{\link{within_group}},
  \code{\link{within_node}}
}

