% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{n}
\alias{n}
\title{Number of items in current group}
\usage{
n()
}
\value{
Number of items in current group (or node if ungrouped)
}
\description{
Number of items in current group
}
\examples{
\donttest{
dat <- Multiplyr (x=1:100, G=rep(1:4, each=25), cl=2)
dat \%>\% group_by (G)
dat \%>\% summarise (G.size = n())
dat \%>\% shutdown()
}
}
\seealso{
Other utility functions: \code{\link{between}},
  \code{\link{cumall}}, \code{\link{cumany}},
  \code{\link{cummean}}, \code{\link{distribute}},
  \code{\link{first}}, \code{\link{lag}},
  \code{\link{last}}, \code{\link{lead}},
  \code{\link{n_distinct}}, \code{\link{n_groups}},
  \code{\link{nth}}
}

