% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{partition_group}
\alias{partition_group}
\alias{partition_group_}
\title{Partition data so that each group is wholly on a node}
\usage{
partition_group(.self, ...)

partition_group_(.self, ..., .dots)
}
\arguments{
\item{.self}{Data frame}

\item{...}{Names of grouping variables}

\item{.dots}{Workaround for non-standard evaluation}
}
\description{
Partition data so that each group is wholly on a node
}
\section{Functions}{
\itemize{
\item \code{partition_group_}: 
}}

