% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{rename}
\alias{rename}
\alias{rename_}
\title{Rename variables}
\usage{
rename(.self, ...)

rename_(.self, ..., .dots)
}
\arguments{
\item{.self}{Data frame}

\item{...}{Additional parameters}

\item{.dots}{Workaround for non-standard evaluation}
}
\value{
Data frame
}
\description{
Takes a list of newname=oldname pairs and renames columns
}
\examples{
\donttest{
dat <- Multiplyr (x = 1:100)
dat \%>\% rename(y=x)
dat["y"]
dat \%>\% shutdown()
}
}
\seealso{
Other column manipulations: \code{\link{define}},
  \code{\link{define_}}; \code{\link{select}},
  \code{\link{select_}}; \code{\link{undefine}},
  \code{\link{undefine_}}, \code{\link{unselect}},
  \code{\link{unselect_}}
}

