% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{select}
\alias{select}
\alias{select_}
\title{Retain only specified variables}
\usage{
select(.self, ...)

select_(.self, ..., .dots)
}
\arguments{
\item{.self}{Data frame}

\item{...}{Additional parameters}

\item{.dots}{Workaround for non-standard evaluation}
}
\value{
Data frame
}
\description{
Takes a list of columns and returns a data frame with only those columns and in the order specified
}
\examples{
\donttest{
dat <- Multiplyr (x = 1:100, y = 100:1, z = rep(123, 100))
dat \%>\% select (y, x)
dat \%>\% shutdown()
}
}
\seealso{
Other column manipulations: \code{\link{add_rownames}},
  \code{\link{define}}, \code{\link{rename}},
  \code{\link{undefine}}
}

