% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{summarise}
\alias{summarise}
\alias{summarise_}
\title{Summarise data}
\usage{
summarise(.self, ...)

summarise_(.self, ..., .dots)
}
\arguments{
\item{.self}{Data frame}

\item{...}{Additional parameters}

\item{.dots}{Workaround for non-standard evaluation}
}
\description{
Summarise data
}
\section{Functions}{
\itemize{
\item \code{summarise_}: 
}}

