% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{transmute}
\alias{transmute}
\alias{transmute_}
\title{Change variables and drop all others}
\usage{
transmute(.data, ...)

transmute_(.data, ..., .dots)
}
\arguments{
\item{.data}{Data frame}

\item{...}{Additional parameters}

\item{.dots}{Workaround for non-standard evaluation}
}
\description{
Change variables and drop all others
}
\section{Functions}{
\itemize{
\item \code{transmute_}: 
}}

