% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R, R/ops.R
\name{undefine}
\alias{undefine}
\alias{undefine_}
\alias{unselect}
\alias{unselect_}
\title{Delete variables}
\usage{
undefine(.self, ...)

unselect(.self, ...)

undefine_(.self, ..., .dots)

unselect_(.self, ..., .dots)
}
\arguments{
\item{.self}{Data frame}

\item{...}{Additional parameters}

\item{.dots}{Workaround for non-standard evaluation}
}
\value{
Data frame
}
\description{
In much the same way that \code{define} creates new columns, \code{undefine} will delete them
}
\examples{
\donttest{
dat <- Multiplyr (x=1:100, y=100:1)
dat \%>\% undefine (y)
dat \%>\% shutdown()
}
}
\seealso{
Other column manipulations: \code{\link{add_rownames}},
  \code{\link{define}}, \code{\link{rename}},
  \code{\link{select}}
}

