% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\name{ungroup}
\alias{rowwise}
\alias{ungroup}
\title{Return data to non-grouped}
\usage{
ungroup(.self)

rowwise(.self)
}
\arguments{
\item{.self}{Data frame}
}
\value{
Data frame
}
\description{
After grouping data with group_by, there may be a need to return to a
non-grouped form. Running ungroup() will drop any grouping. This can be
reinstated again with regroup().
}

